step-4 | 🔧 fix duplicate code
